requirements.txt
python-magic
sortedcontainers
